fix misleading indention
most times this was caused by
if (x) free(x); x=0;
this was fixed by removing the if() and by the way setting x=NULL (it is a pointer)

second cause was if (x>0) close(x); x=-1; this is fixed by sending x=-1 to
the next line.

If save i removed same useless lines.